+++ macro/arch.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-
 #include "unused.h"
 
 int main(void)
@@ -7,23 +6,29 @@ int main(void)
 	int __unused *p = NULL;
 
 #if defined(__i386__)
-	printf("x86-32\n");
+	printf("x86-32 ");
 	printf("sizeof(*) = %d\n", sizeof(p));
 #elif defined(__x86_64__)
-	printf("x86-64\n");
+	printf("x86-64 ");
 	printf("sizeof(*) = %ld\n", sizeof(p));
 #elif defined(__aarch64__)
-	printf("aarch64\n");
+	printf("aarch64 ");
 	printf("sizeof(*) = %ld\n", sizeof(p));
 #elif defined(__riscv)
 # if __riscv_xlen == 64
-	printf("riscv64\n");
-	printf("sizeof(*) = %ld\n", sizeof(p));
+	printf("riscv64 ");
 # else
-	printf("riscv32\n");
+	printf("riscv32 ");
 # endif
+	printf("sizeof(*) = %ld\n", sizeof(p));
 #elif defined(__loongarch64)
-	printf("loongarch64\n");
+	printf("loongarch64 ");
+	printf("sizeof(*) = %ld\n", sizeof(p));
+#elif defined(__powerpc64__)
+#if defined(__PPC64__)
+	printf("__PPC64__: ");
+#endif
+	printf("ppc64 ");
 	printf("sizeof(*) = %ld\n", sizeof(p));
 #endif
 	return 0;